Clean up release structure
Split oversized LabRun, TUI, shell, API, streaming, and agent-tool modules by responsibility.

Add release/refinement docs, source line ceiling guard, command maturity labels, and public module boundary cleanup.

Validation: cargo fmt --check; cargo check -q; cargo check --features legacy-cli -q; cargo check --features experimental-api-server -q; cargo doc --no-deps -q; cargo clippy --all-targets --all-features -- -D warnings; cargo test -q; bash scripts/validate_docs.sh; git diff --check

+++ src/lab/context.rs
@@ -41,7 +41,7 @@ impl LabContextLayer {
             layer: layer.into(),
             label: label.into(),
             stability,
-            estimated_tokens: estimate_tokens(&content) as u64,
+            estimated_tokens: estimate_tokens(&content),
             content,
         }
     }
